+++ examples/preferencial_attachment/preferencial_attachment/model.py
@@ -3,16 +3,17 @@ import numpy as np
 from mesa import Model, DataCollector
 from mesa.discrete_space import Network
 
-from agents import NodeAgent
+from .agents import NodeAgent
 
-def calculate_total_degree(model):
-    _ , degree = zip(*model.graph.degree())
-    return sum(degree)
+# def calculate_total_degree(model):
+#     _ , degree = zip(*model.graph.degree())
+#     return sum(degree)
 
 class AgentNetwork(Model):
     
-    def __init__(self,num=100,seed=None):
+    def __init__(self,num=100,seed=42):
         super().__init__(seed=seed)
+        self.random = seed
         self.num = num
         self.graph = nx.Graph()
         self.new_node=1
@@ -21,18 +22,17 @@ class AgentNetwork(Model):
 
         self.graph.add_edge(0,1)
 
-        self.datacollector = DataCollector(
-            {
-            "Degree": calculate_total_degree,
-            }
-        )
+        # self.datacollector = DataCollector(
+        #     {
+        #     "Degree": calculate_total_degree,
+        #     }
+        # )
     
         self.grid = Network(self.graph,capacity=1,random=self.random) 
-        NodeAgent.create_agents(model=self,n=self.num,cell=list(self.grid.all_cells))
+        NodeAgent.create_agents(model=self,n=self.num,cell=list(self.grid.all_cells),)
     
 
     def step(self):
-        print("taking steppp...........")
         self.new_node += 1
 
         # if self.new_node >= self.num:
@@ -41,7 +41,10 @@ class AgentNetwork(Model):
         # self.graph.add_node(self.new_node)
 
         # extract node IDs along with their degrees
+        print(self.graph.degree())
         nodes, degree = zip(*self.graph.degree())
+        print(f"nodes: {nodes}")
+        print(f"degree: {type(degree)}")
         total_degree = sum(degree)
 
         # probabilities of connecting to an existing node
@@ -56,4 +59,4 @@ class AgentNetwork(Model):
         # for x in self.agents:
         #     print(x.unique_id)
 
-        self.datacollector.collect(self)
+        # self.datacollector.collect(self)
